New prefabs, sprite size
- Never change the transform properties of the prefabs. When adding
units or buildings to a level directly drag the prefab onto the
environmentprefab in the level. Then it will be properly placed.
- The row prefabs do not have an sprite by default. When designing a
level the correct sprite and environment type need to be configured. Do
not use the grassenvprefabs and such.
- In the future more sized levelsizeprefabs will be made right now the
only one is 18X10.
- Remember when adding units or buildings onto a tile to drag that
gameobject into the correct gameobject slot in the tile script which is
located on the environmenttemplate within a row.