Unit transform parent
Set the unit transform parent to the new tile transform which is has
moved to. This way it is always perfectly placed onto the tile and its
position is always 0,0,0 in the end. And it moves the unit in the unity
hierarchie under the new tile object as a parent.