Gamebar, IGameloop, small refactors / cleanup
- Removed the old gamebar from the maincamera prefab.
- Gamebar script now only renders the current player and gold in the top
left corner.
- Various classes now implements IGameloop. The gameloop calls all of
these methods.
- Moved some code around in order to make use of the IGameloop
interface.
- Moved all of the camera control code from the gameloop to the
CameraControls class.
- Events for starting and ending an animation (same event.) The attack
class sends an event for starting an animation fight. And when the fight
is over the AnimationInfo class sends an event to the attack class for
calculating the damage and so forth.
- Some small improvements and cleanups. For example you are now unable
to select a new unit when a battle is happening. Capitol letters in the
unit classes.