DayNight turn implementation
- Renamed FogOfWar class to DayStateController
- Added a JSON for Level1 which includes the amount of turns for each
daystate (e.g. morning, midday, evening, night) and the name of the
level and description.
- Added a LevelManager for loading the various scenes in the game.
Changed some code in order to use this.
- Removed the CurrentTurn from the GameManager.
- Added a DayTurnCounter in the DayStateController for checking the
turns the current DayState has been in, and moves the the next DayState.
- Disabled the FOW showing. It is now disabled by default. Will change
this in the future. For now it is always day. Need to discuss how the
various daystates will effect the game in terms of logic and visuals.