Directional light problem fixed
- The light caused sprites that where rotated 180 degress to be very
dark. This is because of how this and the shaders work. So placed a
second light sources on the back shining in the opposite direction in
order to have the correct lightning. Kind of ugly but because of lack of
shader experience this was the easiest. And according to unity the
directional light is the least expansive to calculate.