Fixed endgame bug. See description.
Because the _scripts and stuff gets set to inactive the static
TileHelper methods can no longer access GameManager. This caused null
references. Instead loaded the main menu. A new scene called AfterMath
of ResultScreen or something must be created.  if the game was won then
the rest of the end turn calculating would also continue in the
background, including fading and such. So loading a new scene makes sure
that no unnecessary code is being run.